// backend/migrations/snapshots/7d944574500b879bec801cb46b1e0c5daca3eaffdc0d298537963acfde8969e5/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Product: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly handle: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly descriptionHtml: CodecTypes['pg/text@1']['output'] | null;
      readonly fabric: CodecTypes['pg/text@1']['output'] | null;
      readonly work: CodecTypes['pg/text@1']['output'] | null;
      readonly category: CodecTypes['pg/text@1']['output'] | null;
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly comparePrice: CodecTypes['pg/int4@1']['output'] | null;
      readonly isNewArrival: CodecTypes['pg/bool@1']['output'];
      readonly isBestseller: CodecTypes['pg/bool@1']['output'];
      readonly isSale: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly ProductImage: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly alt: CodecTypes['pg/text@1']['output'] | null;
      readonly productId: CodecTypes['pg/int4@1']['output'];
    };
    readonly ProductVariant: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly sku: CodecTypes['pg/text@1']['output'] | null;
      readonly size: CodecTypes['pg/text@1']['output'] | null;
      readonly color: CodecTypes['pg/text@1']['output'] | null;
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly compareAtPrice: CodecTypes['pg/int4@1']['output'] | null;
      readonly stock: CodecTypes['pg/int4@1']['output'];
      readonly imageUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly productId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Product: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly handle: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly descriptionHtml: CodecTypes['pg/text@1']['input'] | null;
      readonly fabric: CodecTypes['pg/text@1']['input'] | null;
      readonly work: CodecTypes['pg/text@1']['input'] | null;
      readonly category: CodecTypes['pg/text@1']['input'] | null;
      readonly price: CodecTypes['pg/int4@1']['input'];
      readonly comparePrice: CodecTypes['pg/int4@1']['input'] | null;
      readonly isNewArrival: CodecTypes['pg/bool@1']['input'];
      readonly isBestseller: CodecTypes['pg/bool@1']['input'];
      readonly isSale: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly ProductImage: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly alt: CodecTypes['pg/text@1']['input'] | null;
      readonly productId: CodecTypes['pg/int4@1']['input'];
    };
    readonly ProductVariant: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly sku: CodecTypes['pg/text@1']['input'] | null;
      readonly size: CodecTypes['pg/text@1']['input'] | null;
      readonly color: CodecTypes['pg/text@1']['input'] | null;
      readonly price: CodecTypes['pg/int4@1']['input'];
      readonly compareAtPrice: CodecTypes['pg/int4@1']['input'] | null;
      readonly stock: CodecTypes['pg/int4@1']['input'];
      readonly imageUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly productId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly product_images: {
      readonly alt: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
    };
    readonly product_variants: {
      readonly color: CodecTypes['pg/text@1']['output'] | null;
      readonly compareAtPrice: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly imageUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly size: CodecTypes['pg/text@1']['output'] | null;
      readonly sku: CodecTypes['pg/text@1']['output'] | null;
      readonly stock: CodecTypes['pg/int4@1']['output'];
    };
    readonly products: {
      readonly category: CodecTypes['pg/text@1']['output'] | null;
      readonly comparePrice: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly descriptionHtml: CodecTypes['pg/text@1']['output'] | null;
      readonly fabric: CodecTypes['pg/text@1']['output'] | null;
      readonly handle: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isBestseller: CodecTypes['pg/bool@1']['output'];
      readonly isNewArrival: CodecTypes['pg/bool@1']['output'];
      readonly isSale: CodecTypes['pg/bool@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly work: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly product_images: {
      readonly alt: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
    };
    readonly product_variants: {
      readonly color: CodecTypes['pg/text@1']['input'] | null;
      readonly compareAtPrice: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly imageUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly price: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly size: CodecTypes['pg/text@1']['input'] | null;
      readonly sku: CodecTypes['pg/text@1']['input'] | null;
      readonly stock: CodecTypes['pg/int4@1']['input'];
    };
    readonly products: {
      readonly category: CodecTypes['pg/text@1']['input'] | null;
      readonly comparePrice: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly descriptionHtml: CodecTypes['pg/text@1']['input'] | null;
      readonly fabric: CodecTypes['pg/text@1']['input'] | null;
      readonly handle: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly isBestseller: CodecTypes['pg/bool@1']['input'];
      readonly isNewArrival: CodecTypes['pg/bool@1']['input'];
      readonly isSale: CodecTypes['pg/bool@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly price: CodecTypes['pg/int4@1']['input'];
      readonly work: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;